requirements.txt
absl-py==0.7.1
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
dill==0.3.0
future==0.17.1
gast==0.2.2
GPUtil==1.4.0
grpcio==1.20.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.1.1
more-itertools==7.0.0
networkx==2.3
numpy==1.17.0
opencv-python==4.1.0.25
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.11.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.0
pyparsing==2.4.0
PyPDF2==1.26.0
pytest==4.4.2
python-dateutil==2.8.0
pytorch-msssim==0.1.3
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
requests==2.21.0
scikit-image==0.15.0
scikit-learn==0.21.0
scipy==1.3.0rc2
seaborn==0.9.0
shap==0.28.5
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.2
